chore: rename sentry-tunnel to sentryTunnel
